requirements.txt
youtube_dl
colorama==0.3.7
psutil==4.1.0
aiohttp==0.21.5
beautifulsoup4==4.4.1
blessings==1.6
chardet==2.3.0
curtsies==0.2.6
decorator==4.0.9
discord.py==0.10.0a0
docopt==0.6.2
google==1.9.1
greenlet==0.4.9
monotonic==1.0
praw==3.4.0
Pygments==2.1.3
pyowm==2.3.1
requests==2.9.1
six==1.10.0
update-checker==0.11
wcwidth==0.1.6
websockets==2.7
ws4py==0.3.4
youtube-dl==2016.4.5
zipdep.py==1.2.0
google-api-python-client>=1.5.0